+++ kcommons/src/main/java/com/kairos/dto/user/country/agreement/cta/cta_response/AccessGroupDTO.java
@@ -7,6 +7,7 @@ import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.validation.constraints.AssertTrue;
 import javax.validation.constraints.NotNull;
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.Set;
 @Getter
@@ -21,7 +22,7 @@ public class AccessGroupDTO {
     @NotNull(message = "error.startDate.notnull")
     private LocalDate startDate;
     private LocalDate endDate;
-    private Set<Long> dayTypeIds;
+    private Set<BigInteger> dayTypeIds;
     private boolean allowedDayTypes;
 
     @AssertTrue(message = "Access group can't be blank")
